# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Reverse HH API
  version: 0.1.0
  description: MVP API — employers respond to candidate resumes
servers:
  - url: /api/v1
    description: Same-origin API
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        is_admin:
          type: boolean
        has_candidate_profile:
          type: boolean
        companies:
          type: array
          items:
            $ref: '#/components/schemas/CompanyMembership'
    CompanyMembership:
      type: object
      properties:
        company_id:
          type: string
          format: uuid
        company_name:
          type: string
        role:
          type: string
          enum: [owner, recruiter]
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
    CandidateProfileRequest:
      type: object
      required: [display_name]
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 255
    CandidateProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...
security:
  - cookieAuth: []
paths:
  /auth/register:
    post:
      tags: [auth]
      security: []
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          headers:
            Set-Cookie:
              schema:
                type: string
        '400':
          description: Validation error
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
        '401':
          description: Invalid credentials
...
  /auth/me:
    get:
      tags: [auth]
      summary: Current user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
  /candidate/profile:
    post:
      tags: [resumes]
      summary: Create candidate profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateProfileRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateProfile'
        '400':
          description: Profile already exists
        '401':
          description: Not authenticated
    patch:
      tags: [resumes]
      summary: Update candidate profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateProfileRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateProfile'
        '404':
          description: Profile not found
        '401':
          description: Not authenticated
...